Add solution updating endpoint and tests in solutions app

+++ solutions/views.py
@@ -1,7 +1,7 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from .serializers import SolutionSerializer
+from .serializers import SolutionSerializer, SolutionUpdateSerializer
 from .models import Solution
 
 
@@ -28,14 +28,44 @@ class SolutionAPI(APIView):
         
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
+    def put(self, request):
+        """
+        ### request data
+        - solution id
+        - code
+        - lang
+        - solved
+        - caption
+        """
+        if 'id' not in request.data:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        try:
+            found_solution = Solution.objects.get(id=request.data['id'])
+        except Solution.DoesNotExist:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+
+        serializer = SolutionUpdateSerializer(found_solution, data=request.data)
+        if serializer.is_valid():
+            solution = serializer.save()
+            if solution:
+                new_serializer = SolutionSerializer(solution)
+                return Response(new_serializer.data, status=status.HTTP_200_OK)
+        
+        print(serializer.errors)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        
+
     
     def delete(self, request):
         """
         ### request data
         - solution id
         """
+        if 'id' not in request.data:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
         try:
-            solution = Solution.objects.get(id=request.data['solutionId'])
+            solution = Solution.objects.get(id=request.data['id'])
             solution.delete()
         except Solution.DoesNotExist:
             return Response(status=status.HTTP_400_BAD_REQUEST)
